Moving libraries to common location
