+++ common/slackBot/index.js
