+++ common/snowflakeQuery/index.js
